colisão entre as mumias arrumado! ficou legal agora hehe
git-svn-id: http://xp-dev.com/svn/horus_eye/trunk@244 863e7ce6-8861-447e-8dcb-03987b876380

+++ src/game/utils/geometryprimitives.cc
@@ -16,6 +16,6 @@ double GPdistance(Vector2D a,Vector2D b){
 }
 
 double GPdistance(int ax,int ay,int bx,int by){
-    return sqrt((ax-bx) * (ax-bx) + (ay-by) * (ay-by));
+    return sqrt(double((ax-bx) * (ax-bx) + (ay-by) * (ay-by)));
 }
 }
